Fixed db insertion of shapes

+++ DB/gtfsdb/model/gtfs.py
@@ -5,9 +5,8 @@ import tempfile
 import time
 from urllib import request
 import zipfile
-
-from DB.gtfsdb import config
-from .route import Route
+import os
+from DB import config
 
 log = logging.getLogger(__name__)
 
@@ -30,22 +29,21 @@ class GTFS(object):
         self.file = filename
         self.local_file = request.urlopen(filename)
 
-    def load(self, db, shouldLoadFile=False):
+    def load(self, db):
         '''Load GTFS into database'''
-        start_time = time.time()
-        log.debug('GTFS load data file: {0}'.format(self.file))
-
-        # load known GTFS files, derived tables & lookup tables
-        gtfs_directory = self.unzip(path="tmp", overwrite=True)
-        if shouldLoadFile:
+        if config.SHOULD_LOAD_DATA:
+            start_time = time.time()
+            log.debug('GTFS load data file: {0}'.format(self.file))
+            # load known GTFS files, derived tables & lookup tables
+            gtfs_directory = self.unzip(path="tmp", overwrite=config.SHOULD_OVERWRITE_ZIP_FILES)
             for cls in db.sorted_classes:
                 if cls.__name__ not in class_not_to_load:
                     print("Loading {0}".format(cls.__name__))
                     cls.load(db=db, gtfs_directory=gtfs_directory)
-        shutil.rmtree(gtfs_directory)
-        print("Finished loading classes")
-        process_time = time.time() - start_time
-        log.debug('GTFS.load ({0:.0f} seconds)'.format(process_time))
+            shutil.rmtree(gtfs_directory)
+            print("Finished loading classes")
+            process_time = time.time() - start_time
+            log.debug('GTFS.load ({0:.0f} seconds)'.format(process_time))
 
     def unzip(self, path: str = None, overwrite=False):
         """This function loads the db with that data from a given zip file.
@@ -55,7 +53,6 @@ class GTFS(object):
         """
         path = path if path else tempfile.mkdtemp()
         if overwrite:
-            print("overwrite")
             try:
                 with closing(zipfile.ZipFile(self.local_file)) as z:
                     z.extractall(path)
